Add auto-update option to pangolin
Adds a `--update` option to pangolin which automatically updates
pangolin, pangoLEARN, and lineages to the latest stable release and
exits (or if already the latest just checks and exits).

Specifically:

    - Latest release tags are fetched from the github releases API
    - These are compared to versions of the currently installed (pangolin,
    pangoLEARN, lineages)
    - If there is a new release for any or all of these it updates them
      to that release using pip
    - Once updates are complete or if all packages were at the latest
      release the command just exits with 0 exitcode

As requiring a query file isn't meaningful if a user is just running
`pangolin --update` this adds a small amount of logic to tolerate no
supplied query if `--update` is being used.