Adding UShER as alternative inference engine with new options --usher / --usher-tree.
This adds usher to environment.yml.  If the new options are not used then pangolin operates the same as before.

If --usher-tree <file> is given then pangolin will use usher with the specified protobuf file to assign lineages.
If only --usher is added then pangolin expects to find {data_dir}/lineageTree.pb (alongside the usual pangoLEARN files).

If --threads is added then it will be passed along to usher, but I did not remove the message about --threads being ignored.